// src/app/models/Movie.modal.ts
export interface Movie {
  vote_count: number;
  id: number;
  video: boolean;
  // vote_average: Float32Array;
  title: string;
  // popularity: Float32Array;
  poster_path: string;
  original_language: string;
  original_title: string;
  genre_ids: Array<number>;
  backdrop_path: string;
  adult: boolean;
  overview: string;
  release_date: string;
  }
  